ai/requirements.txt
pillow
twisted
six